third_party/blink/web_tests/platform/generic/editing/selection/caret-at-end-of-text-line-followed-by-block-in-vertical-mode-expected.txt

Some text.

Testcase for bug 104794: Unable to place the caret at the end of the first line, when followed by a block, in the vertical writing mode. To manually verify the issue, try clicking on the empty region of the first vertical line. The caret should be drawn at the end of the text.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS caretRectAtEnd.left is caretRectAtStart.left
PASS caretRectAtEnd.width is caretRectAtStart.width
PASS caretRectAtEnd.height is caretRectAtStart.height
PASS successfullyParsed is true

TEST COMPLETE
